+++ R/lemur.R
@@ -51,12 +51,13 @@ lemur <- function(data, design = ~ 1, col_data = NULL,
   data_mat <- handle_data_parameter(data, on_disk = FALSE, assay = use_assay)
   col_data <- glmGamPoi:::get_col_data(data, col_data)
   if(! all(colnames(col_data) == make.names(colnames(col_data), unique = TRUE))){
-    warning("The column names of 'colData(fit)' are not unique. This ",
-            "can cause problems for downstream analyses.")
+    stop("The column names of 'colData(fit)' are not unique. Please fix this for example by calling: \n",
+            "`colnames(colData(data)) <- make.names(colnames(colData(data)), unique = TRUE)`")
   }
 
   des <- handle_design_parameter(design, data, col_data, verbose = verbose)
   al_des <- des
+  col_data <- des$col_data
 
   if(! is(data, "SummarizedExperiment")){
     data <- SingleCellExperiment(assays = setNames(list(data_mat), use_assay), colData = col_data)